Make deduction formulas

+++ src/javafactura/businessLogic/econSectors/Deductible.java
@@ -2,5 +2,5 @@ package javafactura.businessLogic.econSectors;
 
 public interface Deductible {
 
-    float deduction(float value, boolean interior, int aggregateSize, double coefEmpresa, double coefIndividual);
+    float deduction(float value, boolean interior, int aggregateSize, float coefEmpresa, float coefIndividual);
 }
